Ensure plugins run in sorted-order.
Plugins have names like NN-blah.js which are designed to make
sure that the simpler plugins run first.  To do that we need
to sort them.

Rather than sorting them each time we iterate over the list
of available plugins we update out addition to sort them when
we add them.